BROKEN: Clean up, rewrite README for intended goal
Deleted a lot of README files since the contribution flow of this project is pretty standard.

Rewrote the README with bash instructions for the user guide to be more precise.

However the instructions in the README do not yet work, since this is a big departure from the original project vision.